Fixed R8 bug
Added kapt for moshi codegen hence fixed R8 problem. Also final commit for app launch.
