+++ app/src/main/java/com/octagon_technologies/sky_weather/MainApplication.kt
@@ -15,10 +15,6 @@ class MainApplication: Application() {
     override fun onCreate() {
         super.onCreate()
         if (BuildConfig.DEBUG) Timber.plant(Timber.DebugTree())
-        delayInit()
-    }
-
-    private fun delayInit(){
         applicationScope.launch {
             setRecurringWork()
         }
