+++ app/src/main/java/com/octagon_technologies/sky_weather/repository/FavouriteLocationsRepo.kt
@@ -2,7 +2,7 @@ package com.octagon_technologies.sky_weather.repository
 
 import com.octagon_technologies.sky_weather.repository.database.FavouriteLocationDatabaseClass
 import com.octagon_technologies.sky_weather.repository.database.MainDataBase
-import com.octagon_technologies.sky_weather.repository.network.location.LocationItem
+import com.octagon_technologies.sky_weather.repository.network.location.Location
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import timber.log.Timber
@@ -12,12 +12,12 @@ object FavouriteLocationsRepo {
 
     suspend fun insertFavouriteLocationToLocalStorage(
         mainDataBase: MainDataBase?,
-        locationItem: LocationItem
+        location: Location
     ): Boolean {
         return try {
             withContext(Dispatchers.IO) {
                 mainDataBase?.favouriteLocationDao?.insertFavouriteLocationDatabaseClass(
-                    FavouriteLocationDatabaseClass(locationItem.placeId.toString(), locationItem)
+                    FavouriteLocationDatabaseClass(location.placeId.toString(), location)
                 )
 
                 true
@@ -30,12 +30,12 @@ object FavouriteLocationsRepo {
 
     suspend fun removeFavouriteLocationToLocalStorage(
         mainDataBase: MainDataBase?,
-        locationItem: LocationItem
+        location: Location
     ): Boolean {
         return try {
             withContext(Dispatchers.IO) {
                 mainDataBase?.favouriteLocationDao?.deleteFavouriteLocationDatabaseClass(
-                    FavouriteLocationDatabaseClass(locationItem.placeId.toString(), locationItem)
+                    FavouriteLocationDatabaseClass(location.placeId.toString(), location)
                 )
 
                 true
@@ -48,7 +48,7 @@ object FavouriteLocationsRepo {
 
     suspend fun removeAllFavouriteLocations(
         mainDataBase: MainDataBase?,
-        listOfFavouriteLocations: ArrayList<LocationItem>?
+        listOfFavouriteLocations: ArrayList<Location>?
     ): Boolean {
         return try {
             withContext(Dispatchers.IO) {
@@ -66,7 +66,7 @@ object FavouriteLocationsRepo {
         }
     }
 
-    suspend fun getFavouriteLocationsAsync(mainDataBase: MainDataBase?): ArrayList<LocationItem>? {
+    suspend fun getFavouriteLocationsAsync(mainDataBase: MainDataBase?): ArrayList<Location>? {
         return try {
             withContext(Dispatchers.IO) {
                 val localFavouriteLocations = mainDataBase?.favouriteLocationDao?.getFavouriteLocationsDatabaseClass()?.map {
