+++ app/src/main/java/com/octagon_technologies/sky_weather/repository/network/reverse_geocoding_location/ReverseGeoCodingAddress.kt
@@ -3,6 +3,9 @@ package com.octagon_technologies.sky_weather.repository.network.reverse_geocodin
 
 import com.squareup.moshi.Json
 
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
 data class ReverseGeoCodingAddress(
     @Json(name = "city")
     val city: String?,
