+++ app/src/main/java/com/octagon_technologies/sky_weather/repository/network/selected_daily_forecast/BaroPressure.kt
@@ -1,7 +1,6 @@
-package com.octagon_technologies.sky_weather.network.selected_daily_forecast
+package com.octagon_technologies.sky_weather.repository.network.selected_daily_forecast
 
 
-import com.octagon_technologies.sky_weather.repository.network.selected_daily_forecast.Min
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
