+++ app/src/main/java/com/octagon_technologies/sky_weather/repository/network/single_forecast/CloudBase.kt
@@ -1,5 +1,8 @@
-package com.octagon_technologies.sky_weather.network.single_forecast
+package com.octagon_technologies.sky_weather.repository.network.single_forecast
 
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
 data class CloudBase(
     val units: String?,
     val value: Any?
