+++ app/src/main/java/com/octagon_technologies/sky_weather/repository/network/single_forecast/PrecipitationType.kt
@@ -1,5 +1,8 @@
 package com.octagon_technologies.sky_weather.repository.network.single_forecast
 
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
 data class PrecipitationType(
     val value: String?
 )
