+++ app/src/main/java/com/octagon_technologies/sky_weather/repository/network/single_forecast/WindDirection.kt
@@ -1,5 +1,8 @@
 package com.octagon_technologies.sky_weather.repository.network.single_forecast
 
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
 data class WindDirection(
     val units: String?,
     val value: Double?
