+++ app/src/main/java/com/octagon_technologies/sky_weather/ui/find_location/FindLocationBindingAdapter.kt
@@ -3,26 +3,26 @@ package com.octagon_technologies.sky_weather.ui.find_location
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.octagon_technologies.sky_weather.Theme
-import com.octagon_technologies.sky_weather.repository.network.location.LocationItem
+import com.octagon_technologies.sky_weather.repository.network.location.Location
 import com.octagon_technologies.sky_weather.ui.search_location.each_search_result_item.EachSearchResultItem
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.GroupieViewHolder
 
 fun RecyclerView.getFavouriteLocations(
     theme: Theme?,
-    arrayOfLocationItem: ArrayList<LocationItem>?,
+    arrayOfLocation: ArrayList<Location>?,
     groupAdapter: GroupAdapter<GroupieViewHolder>?,
     lambda: (EachSearchResultItem) -> Unit
 ) {
     layoutManager = LinearLayoutManager(context)
     adapter = groupAdapter
 
-    arrayOfLocationItem?.forEach {
+    arrayOfLocation?.forEach {
         groupAdapter?.add(
             EachSearchResultItem(
                 theme = theme,
                 isLikedByUser = true,
-                locationItem = it,
+                location = it,
                 lambda = lambda
             )
         )
@@ -31,19 +31,19 @@ fun RecyclerView.getFavouriteLocations(
 
 fun RecyclerView.getRecentLocations(
     theme: Theme?,
-    arrayOfLocationItem: ArrayList<LocationItem>?,
+    arrayOfLocation: ArrayList<Location>?,
     groupAdapter: GroupAdapter<GroupieViewHolder>?,
     lambda: (EachSearchResultItem) -> Unit
 ) {
     layoutManager = LinearLayoutManager(context)
     adapter = groupAdapter
 
-    arrayOfLocationItem?.forEach {
+    arrayOfLocation?.forEach {
         groupAdapter?.add(
             EachSearchResultItem(
                 theme = theme,
                 isLikedByUser = true,
-                locationItem = it,
+                location = it,
                 lambda = lambda
             )
         )
