+++ app/src/main/java/com/octagon_technologies/sky_weather/ui/search_location/SearchLocationBindingAdapter.kt
@@ -3,16 +3,16 @@ package com.octagon_technologies.sky_weather.ui.search_location
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.octagon_technologies.sky_weather.Theme
-import com.octagon_technologies.sky_weather.repository.network.location.LocationItem
+import com.octagon_technologies.sky_weather.repository.network.location.Location
 import com.octagon_technologies.sky_weather.ui.search_location.each_search_result_item.EachSearchResultItem
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.GroupieViewHolder
 
 fun RecyclerView.addLocationSuggestionsToRecyclerView(
     theme: Theme?,
-    mapOfLocationItems: Map<String?, LocationItem>,
+    mapOfLocationItems: Map<String?, Location>,
     groupAdapter: GroupAdapter<GroupieViewHolder>,
-    favouriteItemsMap: Map<String?, LocationItem>,
+    favouriteItemsMap: Map<String?, Location>,
     addToFavourite: (EachSearchResultItem) -> Unit
 ) {
     adapter = groupAdapter
@@ -24,7 +24,7 @@ fun RecyclerView.addLocationSuggestionsToRecyclerView(
             val eachSearchResultItem = EachSearchResultItem(
                 theme = theme,
                 isLikedByUser = it in favouriteItemsMap.values,
-                locationItem = it,
+                location = it,
                 lambda = addToFavourite
             )
             groupAdapter.add(eachSearchResultItem)
