+++ app/src/main/java/com/octagon_technologies/sky_weather/ui/search_location/SearchLocationUtils.kt
@@ -1,10 +1,10 @@
 package com.octagon_technologies.sky_weather.ui.search_location
 
-import com.octagon_technologies.sky_weather.repository.network.location.LocationItem
+import com.octagon_technologies.sky_weather.repository.network.location.Location
 import com.octagon_technologies.sky_weather.repository.network.reverse_geocoding_location.ReverseGeoCodingAddress
 import com.octagon_technologies.sky_weather.repository.network.reverse_geocoding_location.ReverseGeoCodingLocation
 
-fun LocationItem.toReverseGeoCodingLocation(): ReverseGeoCodingLocation {
+fun Location.toReverseGeoCodingLocation(): ReverseGeoCodingLocation {
     val reverseGeoCodingAddress = with(address!!) {
         ReverseGeoCodingAddress(city, country, countryCode, postcode, null, state, suburb)
     }
