+++ app/src/main/java/com/octagon_technologies/sky_weather/ui/search_location/each_search_result_item/EachSearchResultItem.kt
@@ -5,14 +5,14 @@ import android.widget.ImageView
 import com.octagon_technologies.sky_weather.R
 import com.octagon_technologies.sky_weather.Theme
 import com.octagon_technologies.sky_weather.databinding.EachSearchResultItemBinding
-import com.octagon_technologies.sky_weather.repository.network.location.LocationItem
+import com.octagon_technologies.sky_weather.repository.network.location.Location
 import com.xwray.groupie.databinding.BindableItem
 import java.util.*
 
 class EachSearchResultItem(
     val theme: Theme?,
     var isLikedByUser: Boolean = false,
-    var locationItem: LocationItem,
+    var location: Location,
     val lambda: (EachSearchResultItem) -> Unit
 ) : BindableItem<EachSearchResultItemBinding>() {
 
@@ -22,7 +22,7 @@ class EachSearchResultItem(
     override fun bind(binding: EachSearchResultItemBinding, position: Int) {
         actualPosition = position
 
-        locationItem.apply {
+        location.apply {
             binding.locationItem = this
             binding.theme = theme
